Rename src.core.parse as src.plugins
