+++ src/plugins/csv/mohela.py
@@ -2,9 +2,9 @@ from datetime import datetime
 
 from loguru import logger
 
-from ..interfaces import IParser
-from ..utils import convert_amount_to_float
-from ..validation import Account, Statement, Transaction
+from core.interfaces import IParser
+from core.utils import convert_amount_to_float
+from core.validation import Account, Statement, Transaction
 
 
 class Parser(IParser):
