+++ src/plugins/csv/occuauto.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from ..utils import convert_amount_to_float
+from core.utils import convert_amount_to_float
 
 
 def get_account_number(array: list[list[str]]) -> str:
