+++ src/plugins/pdf/fidelityhsa.py
@@ -1,17 +1,23 @@
 import re
-from datetime import datetime, timedelta
+from datetime import datetime
 
 from loguru import logger
 
-from ..interfaces import IParser
-from ..utils import PDFReader, convert_amount_to_float, find_param_in_line
-from ..validation import Account, Statement, Transaction
+from core.interfaces import IParser
+from core.utils import (
+    PDFReader,
+    convert_amount_to_float,
+    find_line_startswith,
+    find_param_in_line,
+    find_regex_in_line,
+)
+from core.validation import Account, Statement, Transaction
 
 
 class Parser(IParser):
-    STATEMENT_TYPE = "Wells Fargo Personal Loan"
-    HEADER_DATE = r"%m/%d/%y"
-    LEADING_DATE = re.compile(r"^\d{2}/\d{2}/\d{2}\s")
+    STATEMENT_TYPE = "Fidelity Health Savings Account"
+    HEADER_DATE = r"%B %d, %Y"
+    DATE_REGEX = re.compile(r"[ADFJMNOS]\w*\s\d{1,2}\,\s\d{4}\s-\s")
 
     def parse(self, reader: PDFReader) -> Statement:
         """Entry point
@@ -23,7 +29,6 @@ class Parser(IParser):
             Statement: Statement dataclass
         """
         logger.trace(f"Parsing {self.STATEMENT_TYPE} statement")
-
         try:
             self.lines = reader.extract_lines_simple()
             if not self.lines:
@@ -56,25 +61,26 @@ class Parser(IParser):
         """
         Parse the statement date range into datetime.
 
+        Example:
+        Your Account Summary Statement Period: 10/01/2018 to 10/31/2018
+
         Raises:
             ValueError: If dates cannot be parsed or are invalid.
         """
-        logger.trace("Attempting to parse dates from text.")
         try:
-            pattern = re.compile(r"Statement Date (\d{2}/\d{2}/\d{2})")
-            end_date_str = pattern.search(self.reader.text_simple).group(1)
-            self.end_date = datetime.strptime(end_date_str, self.HEADER_DATE)
+            logger.trace("Attempting to parse dates from text.")
 
-            # No start date or number of days in statement are given in the text.
-            # Use 31 days to ensure coverage calculations still work
-            self.start_date = self.end_date - timedelta(days=31)
+            _, date_line, _ = find_regex_in_line(self.lines, self.DATE_REGEX)
+            date_strs = [word.strip() for word in date_line.split("-")]
+            self.start_date = datetime.strptime(date_strs[0], self.HEADER_DATE)
+            self.end_date = datetime.strptime(date_strs[1], self.HEADER_DATE)
         except Exception as e:
             logger.trace(f"Failed to parse dates from text: {e}")
             raise ValueError(f"Failed to parse statement dates: {e}")
 
     def extract_accounts(self) -> list[Account]:
         """
-        One account per statement
+        One account per Fidelity HSA statement
 
         Returns:
             list[Account]: List of accounts for this statement.
@@ -84,6 +90,10 @@ class Parser(IParser):
     def extract_account(self) -> Account:
         """
         Extracts account-level data, including balances and transactions.
+
+        Args:
+            account_num (str): The account number.
+            lines (list[str]): The lines of text corresponding to the account section.
 
         Returns:
             Account: The extracted account as a dataclass instance.
@@ -99,7 +109,7 @@ class Parser(IParser):
 
         # Extract statement balances
         try:
-            self.start_balance, self.end_balance = self.get_statement_balances()
+            start_balance, end_balance, i_start, i_end = self.get_statement_balances()
         except Exception as e:
             raise ValueError(
                 f"Failed to extract balances for account {account_num}: {e}"
@@ -107,7 +117,7 @@ class Parser(IParser):
 
         # Extract transaction lines
         try:
-            transaction_lines = self.get_transaction_lines()
+            transaction_lines = self.get_transaction_lines(i_start, i_end)
         except Exception as e:
             raise ValueError(
                 f"Failed to extract transactions for account {account_num}: {e}"
@@ -124,38 +134,41 @@ class Parser(IParser):
         # Return the Account dataclass
         return Account(
             account_num=account_num,
-            start_balance=self.start_balance,
-            end_balance=self.end_balance,
+            start_balance=start_balance,
+            end_balance=end_balance,
             transactions=transactions,
         )
 
     def extract_account_number(self) -> str:
-        """Retrieve the account  number from the statement
-
-        Returns:
-            str: Account number
         """
-        pattern = "Account Number "
-        _, line = find_param_in_line(self.lines, pattern)
-        rline = line.split(pattern)[-1]
-        return rline.split()[0].strip()
+        Fidelity 401k statements don't have an account number.
+        Instead retrieve the statement description (e.g. "Company 401k").
+        First line that contains the word '401(k)' is assumed as account number.
+        """
+        search_str = "Account Number:"
+        _, line = find_param_in_line(self.lines, search_str)
+        rline = line.split(search_str)[-1]
+        account_num = rline.split()[0]
+        return account_num
 
     def get_statement_balances(self) -> tuple[float, float]:
         """
-        Extract the starting balance from the statement.
+        Extract the starting and ending balance from the statement.
 
         Raises:
             ValueError: Unable to extract both balances.
         """
-        patterns = ["Prior Principal Balance", "Ending Principal Balance"]
+        patterns = ["Beginning Account Value", "Ending Account Value"]
         balances = {}
+        indices = {}
 
         for pattern in patterns:
             try:
-                _, balance_line = find_param_in_line(self.lines, pattern)
-                balance_str = balance_line.split(pattern)[-1].strip().split()[0]
-                balance_str = balance_str.replace("*", "")
-                balances[pattern] = -convert_amount_to_float(balance_str)
+                i_line, balance_line = find_line_startswith(self.lines, pattern)
+                balance_str = balance_line.split()[-2]
+                balance_str = "0" if balance_str == "-" else balance_str
+                indices[pattern] = i_line
+                balances[pattern] = convert_amount_to_float(balance_str)
                 logger.trace(f"Extracted {pattern}: {balances[pattern]}")
             except ValueError as e:
                 logger.warning(
@@ -163,29 +176,27 @@ class Parser(IParser):
                 )
 
         # Ensure both balances are found
-        if len(balances) != len(patterns):
-            missing = [p for p in patterns if p not in balances]
+        missing = [p for p in patterns if p not in balances]
+        if missing:
             raise ValueError(
                 f"Could not extract balances for patterns: {', '.join(missing)}"
             )
 
-        return balances[patterns[0]], balances[patterns[1]]
+        return (
+            balances[patterns[0]],
+            balances[patterns[1]],
+            indices[patterns[0]],
+            indices[patterns[1]],
+        )
 
-    def get_transaction_lines(self) -> list[str]:
+    def get_transaction_lines(self, i_start: int, i_end: int) -> list[str]:
         """
         Extract lines containing transaction information.
 
         Returns:
             list[str]: Processed lines containing dates and amounts for this statement.
         """
-        transaction_lines = []
-        for line in self.lines:
-            if "Ending Principal Balance" in line:
-                break
-            if self.LEADING_DATE.search(line):
-                transaction_lines.append(line)
-
-        return transaction_lines
+        return self.lines[i_start + 1 : i_end]
 
     def parse_transaction_lines(
         self, transaction_lines: list[str]
@@ -208,50 +219,28 @@ class Parser(IParser):
             if len(words) < 3:
                 raise ValueError(f"Invalid transaction line: {line}")
 
-            # Each transaction has mm/dd/yy format
-            posting_date = datetime.strptime(words[0], self.HEADER_DATE)
-
             # Extract amount
+            amount_str = "0" if words[-2] == "-" else words[-2]
             try:
-                amount = convert_amount_to_float(words[-1])
+                amount = convert_amount_to_float(amount_str)
             except ValueError as e:
                 raise ValueError(f"Error parsing amounts in line '{line}': {e}")
 
-            # Get the description
-            desc = " ".join(words[1:-1])
+            # Extract description
+            desc = " ".join(words[:-2]).replace("*", "")
+
+            # Validate description
             if not desc:
                 raise ValueError(f"Missing description in transaction line: {line}")
 
-            if "INTEREST PAYMENT" in desc:
-                # Hidden interest charge to match the interest payment
-                transactions.append(
-                    Transaction(
-                        transaction_date=posting_date,
-                        posting_date=posting_date,
-                        amount=-amount,
-                        desc="INTEREST FEE",
-                    )
-                )
-
             # Create the Transaction object
-            transactions.append(
-                Transaction(
-                    transaction_date=posting_date,
-                    posting_date=posting_date,
-                    amount=amount,
-                    desc=desc,
-                )
+            transaction = Transaction(
+                transaction_date=self.end_date,
+                posting_date=self.end_date,
+                amount=amount,
+                desc=desc,
             )
 
-        # Add loan origination transactions
-        if len(transactions) == 0:
-            transactions.append(
-                Transaction(
-                    transaction_date=self.end_date,
-                    posting_date=self.end_date,
-                    amount=0.0,
-                    desc="LOAN ORIGINATION",
-                )
-            )
+            transactions.append(transaction)
 
         return transactions
