+++ src/plugins/pdf/usbank.py
@@ -1,21 +1,22 @@
 import re
-from datetime import datetime, timedelta
+from datetime import datetime
 
 from loguru import logger
 
-from ..interfaces import IParser
-from ..utils import (
+from core.interfaces import IParser
+from core.utils import (
     PDFReader,
     convert_amount_to_float,
     find_line_startswith,
+    find_param_in_line,
     get_absolute_date,
 )
-from ..validation import Account, Statement, Transaction
+from core.validation import Account, Statement, Transaction
 
 
 class Parser(IParser):
-    STATEMENT_TYPE = "OCCU NICE Credit Card"
-    HEADER_DATE = r"%m/%d/%y"
+    STATEMENT_TYPE = "US Bank Credit Card"
+    HEADER_DATE = r"%m/%d/%Y"
     DATE_REGEX = re.compile(r"\d{2}/\d{2}")
     LEADING_DATE = re.compile(r"^\d{2}/\d{2}\s")
 
@@ -29,18 +30,32 @@ class Parser(IParser):
             Statement: Statement dataclass
         """
         logger.trace(f"Parsing {self.STATEMENT_TYPE} statement")
-
+        self.reader = reader
         try:
-            # OCCU NICE statements require layout text extraction
-            self.lines = reader.extract_lines_clean()
+            self.lines_simple = reader.extract_lines_simple()
+            self.lines = self.extract_lines_from_reader()
             if not self.lines:
                 raise ValueError("No lines extracted from the PDF.")
-
-            self.reader = reader
             return self.extract_statement()
         except Exception as e:
             logger.error(f"Error parsing {self.STATEMENT_TYPE} statement: {e}")
             raise
+
+    def extract_lines_from_reader(self) -> list[str]:
+        """Uses custom pdfplumber layout tolerances to exclude
+        QR code embedded in statement so metadata extraction is possible
+
+        Returns:
+            list[str]: Whitespace normalized text extracted from pdf
+        """
+        pages = [
+            page.extract_text(layout=True, x_tolerance=0.5, y_tolerance=0.5) or ""
+            for page in self.reader.PDF.pages
+        ]
+        text = "\n".join(pages)
+        lines_raw = [line for line in text.splitlines() if line.strip()]
+        lines_clean = [" ".join(line.split()) for line in lines_raw]
+        return lines_clean
 
     def extract_statement(self) -> Statement:
         """Extracts all statement data
@@ -48,7 +63,7 @@ class Parser(IParser):
         Returns:
             Statement: Statement dataclass
         """
-        self.get_statement_dates()
+        self.get_statement_metadata()
         accounts = self.extract_accounts()
         if not accounts:
             raise ValueError("No accounts were extracted from the statement.")
@@ -59,31 +74,40 @@ class Parser(IParser):
             accounts=accounts,
         )
 
-    def get_statement_dates(self) -> None:
+    def get_statement_metadata(self) -> None:
         """
         Parse the statement date range into datetime.
+        Must use lines extracted via custom layout tolerance
 
         Raises:
             ValueError: If dates cannot be parsed or are invalid.
+
+        Notes:
+            self.start_date, self.end_date, and self.account_num are set
         """
         logger.trace("Attempting to parse dates from text.")
+        patterns = {
+            "start_date": "Open Date:",
+            "end_date": "Closing Date:",
+            "account_num": "Account:",
+        }
+        values = {}
         try:
-            close_str = "StatementClosingDate"
-            ndays_str = "DaysinBillingCycle"
-            _, close_line = find_line_startswith(self.lines, close_str)
-            _, ndays_line = find_line_startswith(self.lines, ndays_str)
+            _, info_line = find_param_in_line(self.lines, patterns["start_date"])
+            for key, pattern in patterns.items():
+                right_str = info_line.split(pattern)[-1].strip()
+                parts = right_str.split()
+                if key == "account_num":
+                    values[key] = "".join(parts)
+                else:
+                    values[key] = parts[0]
 
-            close_line_r = close_line.split(close_str)[-1]
-            ndays_line_r = ndays_line.split(ndays_str)[-1]
-            end_date_str = close_line_r.split()[0]
-            ndays_str = ndays_line_r.split()[0]
-            ndays = int(ndays_str)
-
-            self.end_date = datetime.strptime(end_date_str, self.HEADER_DATE)
-            self.start_date = self.end_date - timedelta(days=ndays - 1)
+            self.start_date = datetime.strptime(values["start_date"], self.HEADER_DATE)
+            self.end_date = datetime.strptime(values["end_date"], self.HEADER_DATE)
+            self.account_num = values["account_num"]
         except Exception as e:
-            logger.trace(f"Failed to parse dates from text: {e}")
-            raise ValueError(f"Failed to parse statement dates: {e}")
+            logger.trace(f"Failed to parse metadata from text: {e}")
+            raise ValueError(f"Failed to metadata statement dates: {e}")
 
     def extract_accounts(self) -> list[Account]:
         """
@@ -104,18 +128,15 @@ class Parser(IParser):
         Raises:
             ValueError: If account number is invalid or data extraction fails.
         """
-        # Extract account number
-        try:
-            account_num = self.extract_account_number()
-        except Exception as e:
-            raise ValueError(f"Failed to extract account number: {e}")
+        # Account number is extracted via self.get_statement_metadata before this
+        # account_num = self.extract_account_number()
 
         # Extract statement balances
         try:
             start_balance, end_balance = self.get_statement_balances()
         except Exception as e:
             raise ValueError(
-                f"Failed to extract balances for account {account_num}: {e}"
+                f"Failed to extract balances for account {self.account_num}: {e}"
             )
 
         # Extract transaction lines
@@ -123,7 +144,7 @@ class Parser(IParser):
             transaction_lines = self.get_transaction_lines()
         except Exception as e:
             raise ValueError(
-                f"Failed to extract transactions for account {account_num}: {e}"
+                f"Failed to extract transactions for account {self.account_num}: {e}"
             )
 
         # Parse transactions
@@ -131,35 +152,25 @@ class Parser(IParser):
             transactions = self.parse_transaction_lines(transaction_lines)
         except Exception as e:
             raise ValueError(
-                f"Failed to parse transactions for account {account_num}: {e}"
-            )
-
-        # Append any interest charges to transactions
-        try:
-            transactions.append(self.interest_transaction())
-        except Exception as e:
-            raise ValueError(
-                f"Failed to parse interest charge for account {account_num}: {e}"
+                f"Failed to parse transactions for account {self.account_num}: {e}"
             )
 
         # Return the Account dataclass
         return Account(
-            account_num=account_num,
+            account_num=self.account_num,
             start_balance=start_balance,
             end_balance=end_balance,
             transactions=transactions,
         )
 
     def extract_account_number(self) -> str:
-        """Retrieve the account  number from the statement
+        """Retrieve the account number from the statement
+        This is handled in get_statement_metadata() for this parser
 
         Returns:
             str: Account number
         """
-        search_str = "AccountNumber "
-        _, line = find_line_startswith(self.lines, search_str)
-        rline = line.split(search_str)[-1]
-        return rline.split()[0]
+        pass
 
     def get_statement_balances(self) -> tuple[float, float]:
         """
@@ -168,13 +179,13 @@ class Parser(IParser):
         Raises:
             ValueError: Unable to extract both balances.
         """
-        patterns = ["PreviousBalance", "NewBalance"]
+        patterns = ["Previous Balance", "New Balance"]
         balances = {}
 
         for pattern in patterns:
             try:
                 _, balance_line = find_line_startswith(self.lines, pattern)
-                balance_str = balance_line.split(pattern)[-1].strip().split()[0]
+                balance_str = balance_line.split()[-1]
                 balances[pattern] = -convert_amount_to_float(balance_str)
                 logger.trace(f"Extracted {pattern}: {balances[pattern]}")
             except ValueError as e:
@@ -194,12 +205,15 @@ class Parser(IParser):
     def get_transaction_lines(self) -> list[str]:
         """
         Extract lines containing transaction information.
+        Must use lines extracted via simple method, since it groups the
+        dates, desc, amount, and `CR` credit indicator consistently.
+        USBank statement format changed slightly in 05/2017 making this necessary.
 
         Returns:
             list[str]: Processed lines containing dates and amounts for this statement.
         """
         transaction_lines = []
-        for line in self.lines:
+        for line in self.lines_simple:
             # Check if the line starts with a valid date
             if self.LEADING_DATE.search(line):
                 transaction_lines.append(line)
@@ -263,28 +277,3 @@ class Parser(IParser):
             transactions.append(transaction)
 
         return transactions
-
-    def interest_transaction(self) -> Transaction:
-        """OCCU NICE statements show interest charges in the header
-        instead of their own transaction line.
-
-        Returns:
-            Transaction: Interest charge
-        """
-        pattern = "InterestCharged"
-
-        try:
-            _, amount_line = find_line_startswith(self.lines, pattern)
-            amount_str = amount_line.split()[2]
-            amount = -convert_amount_to_float(amount_str)
-            logger.trace(f"Extracted {pattern}: {amount}")
-        except ValueError as e:
-            logger.warning(f"Failed to extract balance for pattern '{pattern}': {e}")
-            raise ValueError(f"Could not extract balances for pattern: {pattern}")
-
-        return Transaction(
-            transaction_date=self.end_date,
-            posting_date=self.end_date,
-            amount=amount,
-            desc="Interest Charged",
-        )
