+++ src/plugins/pdf/wfploan.py
@@ -1,24 +1,17 @@
 import re
-from datetime import datetime
+from datetime import datetime, timedelta
 
 from loguru import logger
 
-from ..interfaces import IParser
-from ..utils import (
-    PDFReader,
-    convert_amount_to_float,
-    find_line_startswith,
-    find_param_in_line,
-    get_absolute_date,
-)
-from ..validation import Account, Statement, Transaction
+from core.interfaces import IParser
+from core.utils import PDFReader, convert_amount_to_float, find_param_in_line
+from core.validation import Account, Statement, Transaction
 
 
 class Parser(IParser):
-    STATEMENT_TYPE = "US Bank Credit Card"
-    HEADER_DATE = r"%m/%d/%Y"
-    DATE_REGEX = re.compile(r"\d{2}/\d{2}")
-    LEADING_DATE = re.compile(r"^\d{2}/\d{2}\s")
+    STATEMENT_TYPE = "Wells Fargo Personal Loan"
+    HEADER_DATE = r"%m/%d/%y"
+    LEADING_DATE = re.compile(r"^\d{2}/\d{2}/\d{2}\s")
 
     def parse(self, reader: PDFReader) -> Statement:
         """Entry point
@@ -30,32 +23,17 @@ class Parser(IParser):
             Statement: Statement dataclass
         """
         logger.trace(f"Parsing {self.STATEMENT_TYPE} statement")
-        self.reader = reader
+
         try:
-            self.lines_simple = reader.extract_lines_simple()
-            self.lines = self.extract_lines_from_reader()
+            self.lines = reader.extract_lines_simple()
             if not self.lines:
                 raise ValueError("No lines extracted from the PDF.")
+
+            self.reader = reader
             return self.extract_statement()
         except Exception as e:
             logger.error(f"Error parsing {self.STATEMENT_TYPE} statement: {e}")
             raise
-
-    def extract_lines_from_reader(self) -> list[str]:
-        """Uses custom pdfplumber layout tolerances to exclude
-        QR code embedded in statement so metadata extraction is possible
-
-        Returns:
-            list[str]: Whitespace normalized text extracted from pdf
-        """
-        pages = [
-            page.extract_text(layout=True, x_tolerance=0.5, y_tolerance=0.5) or ""
-            for page in self.reader.PDF.pages
-        ]
-        text = "\n".join(pages)
-        lines_raw = [line for line in text.splitlines() if line.strip()]
-        lines_clean = [" ".join(line.split()) for line in lines_raw]
-        return lines_clean
 
     def extract_statement(self) -> Statement:
         """Extracts all statement data
@@ -63,7 +41,7 @@ class Parser(IParser):
         Returns:
             Statement: Statement dataclass
         """
-        self.get_statement_metadata()
+        self.get_statement_dates()
         accounts = self.extract_accounts()
         if not accounts:
             raise ValueError("No accounts were extracted from the statement.")
@@ -74,40 +52,25 @@ class Parser(IParser):
             accounts=accounts,
         )
 
-    def get_statement_metadata(self) -> None:
+    def get_statement_dates(self) -> None:
         """
         Parse the statement date range into datetime.
-        Must use lines extracted via custom layout tolerance
 
         Raises:
             ValueError: If dates cannot be parsed or are invalid.
-
-        Notes:
-            self.start_date, self.end_date, and self.account_num are set
         """
         logger.trace("Attempting to parse dates from text.")
-        patterns = {
-            "start_date": "Open Date:",
-            "end_date": "Closing Date:",
-            "account_num": "Account:",
-        }
-        values = {}
         try:
-            _, info_line = find_param_in_line(self.lines, patterns["start_date"])
-            for key, pattern in patterns.items():
-                right_str = info_line.split(pattern)[-1].strip()
-                parts = right_str.split()
-                if key == "account_num":
-                    values[key] = "".join(parts)
-                else:
-                    values[key] = parts[0]
+            pattern = re.compile(r"Statement Date (\d{2}/\d{2}/\d{2})")
+            end_date_str = pattern.search(self.reader.text_simple).group(1)
+            self.end_date = datetime.strptime(end_date_str, self.HEADER_DATE)
 
-            self.start_date = datetime.strptime(values["start_date"], self.HEADER_DATE)
-            self.end_date = datetime.strptime(values["end_date"], self.HEADER_DATE)
-            self.account_num = values["account_num"]
+            # No start date or number of days in statement are given in the text.
+            # Use 31 days to ensure coverage calculations still work
+            self.start_date = self.end_date - timedelta(days=31)
         except Exception as e:
-            logger.trace(f"Failed to parse metadata from text: {e}")
-            raise ValueError(f"Failed to metadata statement dates: {e}")
+            logger.trace(f"Failed to parse dates from text: {e}")
+            raise ValueError(f"Failed to parse statement dates: {e}")
 
     def extract_accounts(self) -> list[Account]:
         """
@@ -128,15 +91,18 @@ class Parser(IParser):
         Raises:
             ValueError: If account number is invalid or data extraction fails.
         """
-        # Account number is extracted via self.get_statement_metadata before this
-        # account_num = self.extract_account_number()
+        # Extract account number
+        try:
+            account_num = self.extract_account_number()
+        except Exception as e:
+            raise ValueError(f"Failed to extract account number: {e}")
 
         # Extract statement balances
         try:
-            start_balance, end_balance = self.get_statement_balances()
+            self.start_balance, self.end_balance = self.get_statement_balances()
         except Exception as e:
             raise ValueError(
-                f"Failed to extract balances for account {self.account_num}: {e}"
+                f"Failed to extract balances for account {account_num}: {e}"
             )
 
         # Extract transaction lines
@@ -144,7 +110,7 @@ class Parser(IParser):
             transaction_lines = self.get_transaction_lines()
         except Exception as e:
             raise ValueError(
-                f"Failed to extract transactions for account {self.account_num}: {e}"
+                f"Failed to extract transactions for account {account_num}: {e}"
             )
 
         # Parse transactions
@@ -152,25 +118,27 @@ class Parser(IParser):
             transactions = self.parse_transaction_lines(transaction_lines)
         except Exception as e:
             raise ValueError(
-                f"Failed to parse transactions for account {self.account_num}: {e}"
+                f"Failed to parse transactions for account {account_num}: {e}"
             )
 
         # Return the Account dataclass
         return Account(
-            account_num=self.account_num,
-            start_balance=start_balance,
-            end_balance=end_balance,
+            account_num=account_num,
+            start_balance=self.start_balance,
+            end_balance=self.end_balance,
             transactions=transactions,
         )
 
     def extract_account_number(self) -> str:
-        """Retrieve the account number from the statement
-        This is handled in get_statement_metadata() for this parser
+        """Retrieve the account  number from the statement
 
         Returns:
             str: Account number
         """
-        pass
+        pattern = "Account Number "
+        _, line = find_param_in_line(self.lines, pattern)
+        rline = line.split(pattern)[-1]
+        return rline.split()[0].strip()
 
     def get_statement_balances(self) -> tuple[float, float]:
         """
@@ -179,13 +147,14 @@ class Parser(IParser):
         Raises:
             ValueError: Unable to extract both balances.
         """
-        patterns = ["Previous Balance", "New Balance"]
+        patterns = ["Prior Principal Balance", "Ending Principal Balance"]
         balances = {}
 
         for pattern in patterns:
             try:
-                _, balance_line = find_line_startswith(self.lines, pattern)
-                balance_str = balance_line.split()[-1]
+                _, balance_line = find_param_in_line(self.lines, pattern)
+                balance_str = balance_line.split(pattern)[-1].strip().split()[0]
+                balance_str = balance_str.replace("*", "")
                 balances[pattern] = -convert_amount_to_float(balance_str)
                 logger.trace(f"Extracted {pattern}: {balances[pattern]}")
             except ValueError as e:
@@ -205,16 +174,14 @@ class Parser(IParser):
     def get_transaction_lines(self) -> list[str]:
         """
         Extract lines containing transaction information.
-        Must use lines extracted via simple method, since it groups the
-        dates, desc, amount, and `CR` credit indicator consistently.
-        USBank statement format changed slightly in 05/2017 making this necessary.
 
         Returns:
             list[str]: Processed lines containing dates and amounts for this statement.
         """
         transaction_lines = []
-        for line in self.lines_simple:
-            # Check if the line starts with a valid date
+        for line in self.lines:
+            if "Ending Principal Balance" in line:
+                break
             if self.LEADING_DATE.search(line):
                 transaction_lines.append(line)
 
@@ -241,39 +208,50 @@ class Parser(IParser):
             if len(words) < 3:
                 raise ValueError(f"Invalid transaction line: {line}")
 
-            # Convert leading mm/dd to full datetime
-            posting_date = get_absolute_date(
-                words.pop(0), self.start_date, self.end_date
-            )
-
-            # If there is a second date, it is the transaction date
-            if words and self.DATE_REGEX.search(words[0]):
-                transaction_date = get_absolute_date(
-                    words.pop(0), self.start_date, self.end_date
-                )
-            else:
-                # The single date is the posting date
-                transaction_date = posting_date
+            # Each transaction has mm/dd/yy format
+            posting_date = datetime.strptime(words[0], self.HEADER_DATE)
 
             # Extract amount
             try:
-                amount = -convert_amount_to_float(words[-1])
+                amount = convert_amount_to_float(words[-1])
             except ValueError as e:
                 raise ValueError(f"Error parsing amounts in line '{line}': {e}")
 
             # Get the description
-            desc = " ".join(words[:-1])
+            desc = " ".join(words[1:-1])
             if not desc:
                 raise ValueError(f"Missing description in transaction line: {line}")
 
+            if "INTEREST PAYMENT" in desc:
+                # Hidden interest charge to match the interest payment
+                transactions.append(
+                    Transaction(
+                        transaction_date=posting_date,
+                        posting_date=posting_date,
+                        amount=-amount,
+                        desc="INTEREST FEE",
+                    )
+                )
+
             # Create the Transaction object
-            transaction = Transaction(
-                transaction_date=transaction_date,
-                posting_date=posting_date,
-                amount=amount,
-                desc=desc,
+            transactions.append(
+                Transaction(
+                    transaction_date=posting_date,
+                    posting_date=posting_date,
+                    amount=amount,
+                    desc=desc,
+                )
             )
 
-            transactions.append(transaction)
+        # Add loan origination transactions
+        if len(transactions) == 0:
+            transactions.append(
+                Transaction(
+                    transaction_date=self.end_date,
+                    posting_date=self.end_date,
+                    amount=0.0,
+                    desc="LOAN ORIGINATION",
+                )
+            )
 
         return transactions
